--- a/.gitignore
+++ b/.gitignore
@@ -1,30 +1,5 @@
-# Prerequisites
-*.d
-
 # Compiled Object files
-*.slo
-*.lo
 *.o
-*.obj
-
-# Precompiled Headers
-*.gch
-*.pch
-
-# Compiled Dynamic libraries
-*.so
-*.dylib
-*.dll
-
-# Fortran module files
-*.mod
-*.smod
-
-# Compiled Static libraries
-*.lai
-*.la
-*.a
-*.lib
 
 # Executables
 *.out
@@ -40,11 +15,3 @@
 
 # MacOS
 *.DS_STORE
-
-# Misc
-*.*
-
-# Versions folder
-versions/
-
-Thrawn