--- a/src/constants.h
+++ b/src/constants.h
@@ -8,6 +8,16 @@
 
 using namespace std;
 
+// ----------------------------------------
+// Some constants used in the search
+// ----------------------------------------
+#define full_depth_moves 4
+#define reduction_limit 3
+#define INFINITY 50000
+#define mateVal 49000
+#define mateScore 48000
+#define MAX_DEPTH 64
+
 enum{
     white,
     black,