--- a/src/globals.cpp
+++ b/src/globals.cpp
@@ -1,0 +1,4 @@
+#include "globals.h"
+
+TranspositionTable tt;
+thrawn::Position pos;