--- a/src/globals.cpp
+++ b/src/globals.cpp
@@ -1,4 +1,6 @@
 #include "globals.h"
+#include "constants.h"
 
 TranspositionTable* tt = new TranspositionTable();
 thrawn::Position* pos = new thrawn::Position();
+ThreadData threadDatas[MAX_THREADS];