--- a/src/globals.h
+++ b/src/globals.h
@@ -1,0 +1,10 @@
+#ifndef GLOBALS_H
+#define GLOBALS_H
+
+#include "transposition_table.h"
+#include "position.h"
+
+extern TranspositionTable tt;
+extern thrawn::Position pos;
+
+#endif // GLOBALS_H