--- a/src/globals.h
+++ b/src/globals.h
@@ -4,7 +4,7 @@
 #include "transposition_table.h"
 #include "position.h"
 
-extern TranspositionTable tt;
-extern thrawn::Position pos;
+extern TranspositionTable* tt;
+extern thrawn::Position* pos;
 
 #endif // GLOBALS_H