--- a/src/globals.h
+++ b/src/globals.h
@@ -2,9 +2,12 @@
 #define GLOBALS_H
 
 #include "transposition_table.h"
+#include  "threading.h"
 #include "position.h"
+#include "constants.h"
 
 extern TranspositionTable* tt;
 extern thrawn::Position* pos;
+extern ThreadData threadDatas[MAX_THREADS];
 
 #endif // GLOBALS_H