--- a/src/main.cpp
+++ b/src/main.cpp
@@ -29,5 +29,6 @@ int main() {
     }
 
     delete pos;
+    delete tt;
     return 0;
 }