--- a/src/perft.h
+++ b/src/perft.h
@@ -6,9 +6,9 @@
 
 extern long leaf_nodes;
 
-void perft_search(thrawn::Position& pos, int depth);
+void perft_search(thrawn::Position* pos, int depth);
 
-int perft_test(thrawn::Position& pos, int depth);
+int perft_test(thrawn::Position* pos, int depth);
 
 void perft_run_unit_tests();