## TLDR
This PR introduces several significant enhancements: , transposition table (TT) handling, and user experience.  
- Introduced Lazy SMP and Multithreading
- Improved Transposition Table handling
- Improved Perft UI + runnable via CLI command
- Added Futility pruning to search

## Key Changes

### Multithreading and Parallel Search (Lazy SMP)
- Implemented Lazy SMP for parallel search across multiple threads.
- Introduced worker threads to run iterative deepening in parallel.
- Introduced `threading.cpp` and `threading.h` to handle ThreadData and parallel search execution.
- Search entry point now supports multiple threads (`search_position_threaded`).
- Uses `std::thread` and atomic counters for synchronization.

### Global Shared Transposition Table (Lockless via XOR Hashing)
- Transposition Table (`TT`) now uses a global shared storage.
- Implemented lockless TT probing using XOR hashing for synchronization-free lookups.
- Age-based replacement logic ensures old entries are efficiently replaced.
- Stored TT entries by encoding depth, score, best move, and hash flag into an unsigned 64bit integer.

### Refactored `Position` Class for Efficient Board State Copying
- Introduced macros for efficient board state saving and restoring.
- `copyBoard(pos)` and `restoreBoard(pos)` macros reduce redundancy in `position.h`.
- Prevents unnecessary deep copies, improving performance during search.

### Improved UI for Perft Testing & CLI Command Support
- Added colored output for better readability.
- Displays progress bars when running perft tests.
- Provides detailed performance ratings based on nodes per second.
- Added `perft` command, allowing perft to be executed via UCI.

### Introduced Futility Pruning for Search
- ~5-10% faster searches

### Increased Stack Size At Compile Time
- Increased stack memory allocation to 8MB for Windows builds.
- Addresses stack overflow issues for precomputed bitboards.

## Benchmarks (Very Preliminary Testing)

When Running with 4 threads compared to Thrawn v2.1 (no parallel search)

| Test Case | Search Depth | v2.1 (time in ms) | Current With 4 Threads (time in ms) | Speed Improvement |
|-----------|-------------|-------------------|-------------------|---------|
| Start Position | 14 | ~8830 | ~2496 | ~3.5x Faster |  
| `3r1rk1/p3p1bp/4bnp1/2p1n3/qp2P3/1N1BBP1N/PPPQ2P1/1K1R3R w - - 8 18` | 12 | ~16779 | ~3307| ~5.1x Faster |